+++ collector/receiver/filereceiver/factory_test.go
