+++ collector/receiver/filereceiver/file_reader_test.go
