+++ collector/receiver/filereceiver/internal/metadata/generated_status.go
