+++ collector/receiver/filereceiver/metadata.yaml
