+++ collector/receiver/filereceiver/receiver_test.go
