+++ collector/receiver/filereceiver/testdata/config.yaml
