+++ collector/receiver/filereceiver/testdata/metrics.json
