+++ collector/receiver/filereceiver/testdata/metrics.pb
